+++ rubik/scores/Score.h
@@ -10,7 +10,7 @@ class Score{
 protected:
 
     /* Score máximo possível */
-    int maxScore = 0;
+    float maxScore = 0;
 
     /* Alvo da pontuação, em relação a qual objeto a pontuação será aplicada */
     const Rubik* target;
@@ -30,6 +30,6 @@ public:
     virtual float calculate(const Rubik& source) = 0;
 
     /* Obtém o score máximo possível */
-    int getMaxScore() const;
+    float getMaxScore() const;
 
 };
